Move SQLite connection fixture to conftest.py for testing and remove redundant code from test files. Ref #147
