+++ tests/core/test_codecs.py
@@ -155,123 +155,6 @@ class TestCodecs:
         with pytest.raises(FileNotFoundError):
             codecs.load('non_existing_file.db')
 
-    @pytest.fixture(name="sqlite3db_connection")
-    def fixture_sqlite3db(self, tmp_path):
-        db_path = tmp_path / "test.db"
-        conn = sqlite3.connect(str(db_path))
-
-        # Create tables and insert test data
-        cursor = conn.cursor()
-        cursor.execute(
-            '''
-            CREATE TABLE gender (
-                gender_id INTEGER PRIMARY KEY,
-                gender TEXT,
-                gender_abbrev TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO gender (gender_id, gender, gender_abbrev) VALUES
-            (1, 'Male', 'M'),
-            (2, 'Female', 'F')
-        '''
-        )
-
-        cursor.execute(
-            '''
-            CREATE TABLE party (
-                party_id INTEGER PRIMARY KEY,
-                party TEXT,
-                party_abbrev TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO party (party_id, party, party_abbrev) VALUES
-            (1, 'Party A', 'PA'),
-            (2, 'Party B', 'PB')
-        '''
-        )
-
-        cursor.execute(
-            '''
-            CREATE TABLE office_type (
-                office_type_id INTEGER PRIMARY KEY,
-                office TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO office_type (office_type_id, office) VALUES
-            (1, 'Office A'),
-            (2, 'Office B')
-        '''
-        )
-
-        cursor.execute(
-            '''
-            CREATE TABLE sub_office_type (
-                sub_office_type_id INTEGER PRIMARY KEY,
-                office_type_id INTEGER,
-                identifier TEXT,
-                description TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO sub_office_type (sub_office_type_id, office_type_id, identifier, description) VALUES
-            (1, 1, 'A', 'Description A'),
-            (2, 2, 'B', 'Description B')
-        '''
-        )
-
-        cursor.execute(
-            '''
-            CREATE TABLE chamber (
-                chamber_id INTEGER PRIMARY KEY,
-                chamber TEXT,
-                chamber_abbrev TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO chamber (chamber_id, chamber, chamber_abbrev) VALUES
-            (1, 'Chamber A', 'CA'),
-            (2, 'Chamber B', 'CB')
-        '''
-        )
-
-        cursor.execute(
-            '''
-            CREATE TABLE government (
-                government_id INTEGER PRIMARY KEY,
-                government TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO government (government_id, government) VALUES
-            (1, 'Government A'),
-            (2, 'Government B')
-        '''
-        )
-
-       
-
-        
-
-        
-
-        conn.commit()
-        return conn
-
 
     def test_codecs_load_with_sqlite_connection(self, sqlite3db_connection):
         codecs = Codecs()
@@ -609,154 +492,7 @@ class TestPersonCodecs:
         with pytest.raises(FileNotFoundError):
             person_codecs.load('non_existing_file.db')
 
-    # FIXME: Use this fixture for all tests. Move to conftest.py
-    @pytest.fixture(name="sqlite3db_connection")
-    def fixture_sqlite3db(self, tmp_path):
-        db_path = tmp_path / "test.db"
-        conn = sqlite3.connect(str(db_path))
-
-        # Create tables and insert test data
-        cursor = conn.cursor()
-        
-        cursor.execute(
-            '''
-            CREATE TABLE gender (
-                gender_id INTEGER PRIMARY KEY,
-                gender TEXT,
-                gender_abbrev TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO gender (gender_id, gender, gender_abbrev) VALUES
-            (1, 'Male', 'M'),
-            (2, 'Female', 'F')
-        '''
-        )
-        
-        cursor.execute(
-            '''
-            CREATE TABLE persons_of_interest (
-                pid INTEGER PRIMARY KEY AUTOINCREMENT,
-                person_id TEXT,
-                name TEXT
-            )
-        '''
-        )
-        
-        cursor.execute(
-            '''
-            INSERT INTO persons_of_interest (person_id, name) VALUES
-            ('p1', 'John Doe'),
-            ('p2', 'Jane Doe')
-        '''
-        )
-        
-        cursor.execute(
-            '''
-            CREATE TABLE chamber (
-                chamber_id INTEGER PRIMARY KEY,
-                chamber TEXT,
-                chamber_abbrev TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO chamber (chamber_id, chamber, chamber_abbrev) VALUES
-            (1, 'Chamber A', 'CA'),
-            (2, 'Chamber B', 'CB')
-        '''
-        )
-
-        cursor.execute(
-            '''
-            CREATE TABLE government (
-                government_id INTEGER PRIMARY KEY,
-                government TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO government (government_id, government) VALUES
-            (1, 'Government A'),
-            (2, 'Government B')
-        '''
-        )
-
-
-        cursor.execute(
-            '''
-            CREATE TABLE office_type (
-                office_type_id INTEGER PRIMARY KEY,
-                office TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO office_type (office_type_id, office) VALUES
-            (1, 'Office A'),
-            (2, 'Office B')
-        '''
-        )
-        
-        cursor.execute(
-            '''
-            CREATE TABLE party (
-                party_id INTEGER PRIMARY KEY,
-                party TEXT,
-                party_abbrev TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO party (party_id, party, party_abbrev) VALUES
-            (1, 'Party A', 'PA'),
-            (2, 'Party B', 'PB')
-        '''
-        )
-        
-        cursor.execute(
-            '''
-            CREATE TABLE sub_office_type (
-                sub_office_type_id INTEGER PRIMARY KEY,
-                office_type_id INTEGER,
-                identifier TEXT,
-                description TEXT
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO sub_office_type (sub_office_type_id, office_type_id, identifier, description) VALUES
-            (1, 1, 'A', 'Description A'),
-            (2, 2, 'B', 'Description B')
-        '''
-        )
-        
-        cursor.execute(
-            '''
-            CREATE TABLE person_party (
-                person_id TEXT,
-                party_id INTEGER,
-                PRIMARY KEY (person_id, party_id)
-            )
-        '''
-        )
-        cursor.execute(
-            '''
-            INSERT INTO person_party (person_id, party_id) VALUES
-            ('p1', 1),
-            ('p2', 2)
-        '''
-        )
-        
-        conn.commit()
-        return conn
+   
 
     @pytest.mark.skip(reason="Fails core/utility.py:121. KeyError: None of ['person_party_id'] are in the columns")
     def test_person_codecs_load_with_existing_file(self, person_codecs, sqlite3db_connection):
